+++ packages/create-svelte/templates/default/src/app.d.ts
@@ -11,6 +11,4 @@ declare namespace App {
 	// interface PrivateEnv {}
 
 	// interface PublicEnv {}
-
-	// interface Session {}
 }