+++ packages/kit/src/runtime/server/index.js
@@ -257,7 +257,7 @@ export async function respond(request, options, state) {
 											event,
 											node,
 											parent: async () => {
-												/** @type {import('types').JSONObject} */
+												/** @type {Record<string, any>} */
 												const data = {};
 												for (let j = 0; j < i; j += 1) {
 													const parent = await promises[j];