+++ packages/kit/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sveltejs/kit
 
+## 1.0.0-next.418
+
+### Patch Changes
+
+- Prevent double import of env modules ([#5955](https://github.com/sveltejs/kit/pull/5955))
+
+* Tighten up params typings, fix load function typings, add event typings to generated types ([#5974](https://github.com/sveltejs/kit/pull/5974))
+
+## 1.0.0-next.417
+
+### Patch Changes
+
+- Render pages without a .svelte file in their proper layout/error files ([#5972](https://github.com/sveltejs/kit/pull/5972))
+
+## 1.0.0-next.416
+
+### Patch Changes
+
+- [fix] allow additional keys in svelte.config.js ([#5961](https://github.com/sveltejs/kit/pull/5961))
+
 ## 1.0.0-next.415
 
 ### Patch Changes